Add Docker image and compose for Frontend (#15)
Containerizes Frontend service to a container via API Gateway at `/`
root and configures a development environment for hot reload on the
website.

Note that working in `frontend` still requires to follow the `README.md`
of the service to `npm install` inside it for development. The Docker
dev image stores the `node_modules` in its image to prevent permission
overriding.